remove codeQL warning about implicit narrowing conversion in compound assignment (#1403)

+++ plugins/examples/rescore/src/main/java/org/opensearch/example/rescore/ExampleRescoreBuilder.java
@@ -222,7 +222,7 @@ public class ExampleRescoreBuilder extends RescorerBuilder<ExampleRescoreBuilder
                                 + "]"
                         );
                     }
-                    topDocs.scoreDocs[i].score *= data.nextValue();
+                    topDocs.scoreDocs[i].score *= (float) data.nextValue();
                 }
             }
             // Sort by score descending, then docID ascending, just like lucene's QueryRescorer
